% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capitulos_br}
\alias{capitulos_br}
\title{Capítulos da R-Ladies no Brasil}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 21 columns.
}
\usage{
capitulos_br
}
\description{
Capítulos da R-Ladies no Brasil
}
\keyword{datasets}
