% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capitulos_redes}
\alias{capitulos_redes}
\title{Redes sociais dos capítulos}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 10 columns.
}
\usage{
capitulos_redes
}
\description{
Redes sociais dos capítulos
}
\keyword{datasets}
