% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eventos_br}
\alias{eventos_br}
\title{Eventos realizados pela R-Ladies no Brasil}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 113 rows and 6 columns.
}
\usage{
eventos_br
}
\description{
Eventos realizados pela R-Ladies no Brasil
}
\keyword{datasets}
