% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eventos_recentes_brasil}
\alias{eventos_recentes_brasil}
\title{Próximos eventos, e eventos recentes, pela R-Ladies no Brasil}
\format{
An object of class \code{data.frame} with 3 rows and 6 columns.
}
\usage{
eventos_recentes_brasil
}
\description{
Próximos eventos, e eventos recentes, pela R-Ladies no Brasil
}
\keyword{datasets}
